another fix

+++ src/libutil/sysutil.cpp
@@ -254,10 +254,10 @@ Sysutil::this_program_path()
     char filename[10240] = "";
 
 #if defined(__linux__)
-    unsigned int size = sizeof(filename);
-    int r             = readlink("/proc/self/exe", filename, size);
+    size_t size = sizeof(filename);
+    ssize_t r = readlink("/proc/self/exe", filename, size);
     // user won't get the right answer if the filename is too long to store
-    OIIO_ASSERT(r < int(size));
+    OIIO_ASSERT(r < ssize_t(size));
     if (r > 0)
         filename[r] = 0;  // readlink does not fill in the 0 byte
 #elif defined(__APPLE__)
